[FIX] Remove some compiler warning and fix coding style

+++ src/common/math/WSymmetricSphericalHarmonic.cpp
@@ -452,7 +452,7 @@ void WSymmetricSphericalHarmonic::normalize()
   double scale = 0.0;
   if ( m_SHCoefficients.size() > 0 )
     scale = std::sqrt( 4.0 * piDouble ) * m_SHCoefficients[0];
-  for ( size_t i = 0; i < m_SHCoefficients.size(); i++ )
+  for ( int i = 0; i < m_SHCoefficients.size(); i++ )
   {
     m_SHCoefficients[ i ] /= scale;
   }
